[add][modify] move getLocalTimeStringRoundUpSecond function to timeFunction.js
